Create checkOwnerUser Decorator
Create a decorator for the owner user checking so this can be used to decorate any method that we only want to give access owner access to.
Save the user object returned from the datastore so we don't have to make another hit later. Add the dictionary from the oauth profile query to
the userData object under .auth so that our methods also have access to that. Import wraps so that __name__ etc get copied accross when being
wrapped by the decorator.