Add TimeZone class
Add a TimeZone database class to act as a look up table for what time zone the locations for a particular day are in. The format mirrors that of the Google TimeZone API.
TimeZone enteries are created with id = time stamp in MS for midday UTC of day
Day is the date,
dstOffset is the offset in seconds for day light saving,
rawOffset is the offset in seconds for the time zone,
timeZoneId is the string ID of the time zone,
timeZoneName is the string name of the time zone.